+++ src/session/rx.c
@@ -92,7 +92,7 @@ int8_t _z_handle_network_message(_z_session_t *zn, _z_zenoh_message_t *msg, uint
                 case _Z_REQUEST_QUERY: {
 #if Z_FEATURE_QUERYABLE == 1
                     _z_msg_query_t *query = &req._body._query;
-                    ret = _z_trigger_queryables(zn, query, req._key, req._rid);
+                    ret = _z_trigger_queryables(zn, query, req._key, (uint32_t)req._rid);
 #else
                     _Z_DEBUG("_Z_REQUEST_QUERY dropped, queryables not supported\n");
 #endif
